Move tests to the root
