+++ __tests__/GlobalStateStore.spec.js
@@ -1,5 +1,5 @@
-var GlobalStateActions = require('../actions/GlobalStateActions');
-var GlobalStateStore = require('../stores/GlobalStateStore');
+var GlobalStateActions = require('../src/js/actions/GlobalStateActions');
+var GlobalStateStore = require('../src/js/stores/GlobalStateStore');
 
 describe('this store', function() {
   it('is can change animating', function() {
