+++ __tests__/levels.spec.js
@@ -2,7 +2,7 @@ var base = require('./base');
 
 describe('GitEngine Levels', function() {
   it('solves levels', function() {
-    var sequences = require('../../levels/index').levelSequences;
+    var sequences = require('../src/levels/index').levelSequences;
     Object.keys(sequences).forEach(function(sequenceKey) {
       var levels = sequences[sequenceKey];
       Object.keys(levels).forEach(function(index) {
