+++ __tests__/simpleRemote.spec.js
